+++ Sources/Core/App/View/VideoPlayer/VideoPlayerView+tvOS.swift
@@ -488,19 +488,19 @@ extension VideoPlayerView {
         onRightArrowTapped?()
     }
 
-    func setControlViewVisibility(visible: Bool, animated: Bool) {
+    func setControlViewVisibility(visible: Bool, withAnimationDuration: Double) {
         DispatchQueue.main.async {
-            UIView.animate(withDuration: animated ? 0.3 : 0) {
+            UIView.animate(withDuration: withAnimationDuration) {
                 self.controlAlphaView.alpha = visible ? 1 : 0
                 self.controlView.alpha = visible ? 1 : 0
             }
         }
     }
 
-    func setInfoViewVisibility(visible: Bool, animated: Bool) {
+    func setInfoViewVisibility(visible: Bool, withAnimationDuration: Double) {
         DispatchQueue.main.async { [weak self] in
             guard let self = self else { return }
-            UIView.animate(withDuration: animated ? 0.2 : 0) {
+            UIView.animate(withDuration: withAnimationDuration) {
                 self.infoView.alpha = visible ? 1 : 0
             }
         }